Python_CNN_Model/requirements.txt
tensorflow>=2.10
numpy
pandas
scikit-learn
Pillow
openpyxl
streamlit
